tiny bit of clean up
