+++ algorithm/test/incubator/SimdFFT.cpp
@@ -1,12 +1,5 @@
 #include "SimdFFT.hpp"
 
-/* detect compiler flavour */
-#if defined(_MSC_VER)
-#define COMPILER_MSVC
-#elif defined(__GNUC__)
-#define COMPILER_GCC
-#endif
-
 #include <algorithm>
 #include <array>
 #include <cassert>
@@ -17,70 +10,6 @@
 #include <memory>
 #include <numbers>
 #include <span>
-
-#if defined(__clang__) || defined(__GNUC__)
-#define ALWAYS_INLINE(return_type) inline return_type __attribute__((always_inline))
-#define NEVER_INLINE(return_type)  return_type __attribute__((noinline))
-#define RESTRICT                   __restrict
-#elif defined(COMPILER_MSVC)
-#define ALWAYS_INLINE(return_type) __forceinline return_type
-#define NEVER_INLINE(return_type)  __declspec(noinline) return_type
-#define RESTRICT                   __restrict
-#endif
-
-#ifdef COMPILER_MSVC
-#pragma warning(disable : 4244 4305 4204 4456)
-#endif
-
-template<typename T>
-constexpr ALWAYS_INLINE(void) store_unchecked(const T& v, typename T::value_type* ptr, auto) {
-    v.copy_to(ptr, stdx::vector_aligned);
-}
-
-template<typename Vec>
-constexpr static ALWAYS_INLINE(auto) blend_lo2_hi1(const Vec& a, const Vec& b) noexcept {
-    // takes low from 2nd arg, high from 1st (compact, FFT convention)
-    // equivalent for lane 'i': out[i] = (i < 2) ? b[i] : a[i];
-    constexpr int size = Vec::size();
-    return vir::simd_permute<size>(stdx::concat(b, a), [](int i) { return i < 2 ? i : i + size; });
-}
-
-template<typename Vec>
-constexpr static ALWAYS_INLINE(auto) interleave(const Vec& in1, const Vec& in2, Vec& out1, Vec& out2) noexcept {
-    constexpr int size   = Vec::size();
-    std::tie(out1, out2) = stdx::split<Vec>(vir::simd_permute(stdx::concat(in1, in2), [](int i) { return (i >> 1) + size * (i & 1); }));
-}
-
-template<typename Vec>
-constexpr static ALWAYS_INLINE(auto) uninterleave(const Vec& in1, const Vec& in2, Vec& out1, Vec& out2) noexcept {
-    constexpr int size   = Vec::size();
-    std::tie(out1, out2) = stdx::split<Vec>(vir::simd_permute(stdx::concat(in1, in2), [](int i) noexcept -> int { return (i % size) * 2 + (i / size); }));
-}
-
-template<typename Vec>
-constexpr static ALWAYS_INLINE(auto) transpose(Vec& x0, Vec& x1, Vec& x2, Vec& x3) noexcept {
-    constexpr int size          = Vec::size();
-    const auto [y0, y1, y2, y3] = stdx::split<Vec>(vir::simd_permute(stdx::concat(x0, x1, x2, x3), [](int i) noexcept -> int { return (i % size) * size + (i / size); }));
-    x0                          = y0;
-    x1                          = y1;
-    x2                          = y2;
-    x3                          = y3;
-}
-
-/* shortcuts for complex multiplications */
-template<typename Vec, typename T>
-constexpr static ALWAYS_INLINE(void) complex_multiply(Vec& ar, Vec& ai, const T& br, const T& bi) noexcept {
-    const Vec tmp = ar * bi;
-    ar            = ar * br - ai * bi;
-    ai            = ai * br + tmp;
-}
-
-template<typename Vec, typename T>
-constexpr static ALWAYS_INLINE(void) complex_multiply_conj(Vec& ar, Vec& ai, const T& br, const T& bi) noexcept {
-    const Vec tmp = ar * bi;
-    ar            = ar * br + ai * bi;
-    ai            = ai * br - tmp;
-}
 
 /**************************************************************************************************************/
 /********************************** complex FFT processing C2C ************************************************/
@@ -1588,90 +1517,6 @@ void unreversed_copy(std::size_t N, const vec<T>* in, vec<T>* out, int out_strid
     uninterleave(h0, g1, out[0], out[1]);
 }
 
-template<Direction direction, std::floating_point T, Transform transform, std::size_t N_>
-constexpr void simdReordering(SimdFFT<T, transform, N_>& setup, std::span<const T> input, std::span<T> output) {
-    using V4                = vec<T, 4>;
-    constexpr std::size_t L = V4::size();
-
-    const std::size_t N     = setup.size();
-    const std::size_t Ncvec = setup.simdVectorSize();
-
-    assert(input.data() != output.data());
-
-    if constexpr (SimdFFT<T, transform, N_>::IsRealValued::value) {
-        assert((SimdFFT<T, transform, N_>::canProcessSize(N, Order::Ordered))); // non-multiple of 32 (while they can be computed) are very hard to bit-reverse
-        if constexpr (direction == Direction::Forward) {
-            const V4* vin  = reinterpret_cast<const V4*>(input.data());
-            V4*       vout = reinterpret_cast<V4*>(output.data());
-
-            const std::size_t dk = N / 32; // For N=48: dk=1
-
-            for (std::size_t k = 0; k < dk; ++k) {
-                V4 out0_0, out0_1, out2_0, out2_1;
-
-                // INTERLEAVE2(vin[k*8+0], vin[k*8+1], vout[2*(0*dk+k)+0], vout[2*(0*dk+k)+1])
-                interleave(vin[k * 8 + 0], vin[k * 8 + 1], out0_0, out0_1);
-                vout[2 * (0 * dk + k) + 0] = out0_0;
-                vout[2 * (0 * dk + k) + 1] = out0_1;
-
-                // INTERLEAVE2(vin[k*8+4], vin[k*8+5], vout[2*(2*dk+k)+0], vout[2*(2*dk+k)+1])
-                interleave(vin[k * 8 + 4], vin[k * 8 + 5], out2_0, out2_1);
-                vout[2 * (2 * dk + k) + 0] = out2_0;
-                vout[2 * (2 * dk + k) + 1] = out2_1;
-            }
-
-            // reversed_copy(dk, vin+2, 8, (v4sf*)(out + N/2))
-            reversed_copy(dk, vin + 2, 8, reinterpret_cast<V4*>(output.data() + N / 2));
-
-            // reversed_copy(dk, vin+6, 8, (v4sf*)(out + N))
-            reversed_copy(dk, vin + 6, 8, reinterpret_cast<V4*>(output.data() + N));
-        } else { // Backward
-            const V4* vin  = reinterpret_cast<const V4*>(input.data());
-            V4*       vout = reinterpret_cast<V4*>(output.data());
-
-            const std::size_t dk = N / 32;
-
-            for (std::size_t k = 0; k < dk; ++k) {
-                V4 out0_0, out0_1, out4_0, out4_1;
-
-                uninterleave(vin[2 * (0 * dk + k) + 0], vin[2 * (0 * dk + k) + 1], out0_0, out0_1);
-                vout[k * 8 + 0] = out0_0;
-                vout[k * 8 + 1] = out0_1;
-
-                uninterleave(vin[2 * (2 * dk + k) + 0], vin[2 * (2 * dk + k) + 1], out4_0, out4_1);
-                vout[k * 8 + 4] = out4_0;
-                vout[k * 8 + 5] = out4_1;
-            }
-
-            unreversed_copy(dk, reinterpret_cast<const V4*>(input.data() + N / 4), reinterpret_cast<V4*>(output.data() + N - 6 * L), -8);
-            unreversed_copy(dk, reinterpret_cast<const V4*>(input.data() + 3 * N / 4), reinterpret_cast<V4*>(output.data() + N - 2 * L), -8);
-        }
-        return;
-    }
-
-    // Complex FFT - this part was already correct
-    const T* inP  = input.data();
-    T*       outP = output.data();
-
-    if constexpr (direction == Direction::Forward) {
-        for (std::size_t k = 0; k < Ncvec; ++k) {
-            const std::size_t kk = (k / 4) + (k % 4) * (Ncvec / 4);
-            V4                lo{}, hi{};
-            interleave(V4(inP + (2 * k + 0) * L, stdx::vector_aligned), V4(inP + (2 * k + 1) * L, stdx::vector_aligned), lo, hi);
-            store_unchecked(lo, outP + (2 * kk + 0) * L, stdx::vector_aligned);
-            store_unchecked(hi, outP + (2 * kk + 1) * L, stdx::vector_aligned);
-        }
-    } else {
-        for (std::size_t k = 0; k < Ncvec; ++k) {
-            const std::size_t kk = (k / 4) + (k % 4) * (Ncvec / 4);
-            V4                re{}, im{};
-            uninterleave(V4(inP + (2 * kk + 0) * L, stdx::vector_aligned), V4(inP + (2 * kk + 1) * L, stdx::vector_aligned), re, im);
-            store_unchecked(re, outP + (2 * k + 0) * L, stdx::vector_aligned);
-            store_unchecked(im, outP + (2 * k + 1) * L, stdx::vector_aligned);
-        }
-    }
-}
-
 template<std::floating_point T>
 constexpr ALWAYS_INLINE(void) realFinalise_4x4(const T* RESTRICT in0, const T* RESTRICT in1, const T* RESTRICT in, std::span<const T> eSpan, T* RESTRICT out) {
     using V                 = vec<T, 4>;
@@ -1874,127 +1719,3 @@ constexpr void realPreprocess(std::size_t Ncvec, std::span<const T> inputSpan, s
     output[(2 * Ncvec - 1) * L + 2] = 2 * (Xi[3] - Xi[1]);
     output[(2 * Ncvec - 1) * L + 3] = -std::numbers::sqrt2_v<T> * (Xr[1] - Xr[3]) - std::numbers::sqrt2_v<T> * (Xi[1] + Xi[3]);
 }
-
-template<Direction direction, Order ordering, std::floating_point T, Transform transform, std::size_t N_>
-constexpr void transformInternal(SimdFFT<T, transform, N_>& setup, std::span<const T> inputSpan, std::span<T> outputSpan, std::span<T> scratch) {
-    using V = SimdFFT<T, transform, N_>::vector_type;
-
-    if constexpr (transform == Transform::Real) {
-        inputSpan  = std::span{inputSpan.data(), setup.size()};
-        outputSpan = std::span{outputSpan.data(), setup.size()};
-    }
-
-    std::span<T>  buff[2]       = {outputSpan, scratch};
-    constexpr int orderinged    = (ordering == Order::Ordered) ? 1 : 0;
-    const int     numStages_odd = setup._radixPlan[1] & 1;
-
-    std::size_t ib = (numStages_odd ^ orderinged) ? 1 : 0;
-
-    // Use Ncvec * 2 for real FFTs (number of vec<T> units), Ncvec for complex
-    constexpr std::size_t L           = V::size();
-    const std::size_t     Ncvec       = setup.simdVectorSize();
-    const std::size_t     n_vecs_real = Ncvec * 2; // Number of vec<T> units for real FFT
-    const std::size_t     n_vecs_cplx = Ncvec;     // Number of vec<T> units for complex FFT
-
-    std::span<const T> twiddle_span = setup.stageTwiddles();
-    std::span<const T> e_span       = setup.butterflyTwiddles();
-
-    if constexpr (direction == Direction::Forward) {
-        ib = !ib;
-        if constexpr (SimdFFT<T, transform, N_>::IsRealValued::value) {
-            std::span<T> outp = rfftf1_ps<T>(n_vecs_real, inputSpan, buff[ib], buff[!ib], twiddle_span, setup._radixPlan);
-            ib                = (outp.data() == buff[0].data()) ? 0 : 1;
-
-            realFinalise<T>(Ncvec, buff[ib], buff[!ib], e_span);
-
-            if constexpr (ordering == Order::Ordered) {
-                // reorder reads from buff[!ib] (where finalize wrote), writes to buff[ib]
-                simdReordering<Direction::Forward>(setup, std::span<const T>{buff[!ib]}, buff[ib]);
-                // ib already points to buff[ib] - correct for final output
-            } else {
-                // For unordered, flip ib to point to buff[!ib] where data is
-                ib = !ib;
-            }
-        } else { // complex path
-            // deinterleave
-            const T* RESTRICT pInput  = std::assume_aligned<64>(inputSpan.data());
-            T* RESTRICT       pBuffer = std::assume_aligned<64>(buff[ib].data());
-            for (std::size_t k = 0; k < Ncvec; ++k) {
-                const std::size_t k2 = 2 * k * L;
-
-                V v0(pInput + k2, stdx::vector_aligned);
-                V v1(pInput + k2 + L, stdx::vector_aligned);
-                V r, i;
-                uninterleave(v0, v1, r, i);
-                store_unchecked(r, pBuffer + k2, stdx::vector_aligned);
-                store_unchecked(i, pBuffer + k2 + L, stdx::vector_aligned);
-            }
-
-            std::span<T> outp = cfftf1_ps<Direction::Backward, T>(n_vecs_cplx, buff[ib], buff[!ib], buff[ib], twiddle_span, setup._radixPlan);
-            ib                = (outp.data() == buff[0].data()) ? 0 : 1;
-
-            complexFinalise(Ncvec, buff[ib], buff[!ib], e_span);
-
-            if constexpr (ordering == Order::Ordered) {
-                simdReordering<Direction::Forward>(setup, std::span<const T>{buff[!ib]}, buff[ib]);
-            } else {
-                ib = !ib;
-            }
-        }
-
-    } else { // Backward direction
-        if (inputSpan.data() == buff[ib].data()) {
-            ib = !ib; // Avoid collision: if input is in current buffer, switch to other
-        }
-
-        if constexpr (ordering == Order::Ordered) {
-            simdReordering<Direction::Backward>(setup, inputSpan, buff[ib]);
-            inputSpan = buff[ib];
-            ib        = !ib;
-        }
-
-        if constexpr (SimdFFT<T, transform, N_>::IsRealValued::value) {
-            realPreprocess<T>(Ncvec, inputSpan, buff[ib], e_span);
-
-            std::span<T> outp = rfftb1_ps<T>(n_vecs_real, buff[ib], buff[0], buff[1], twiddle_span, setup._radixPlan);
-            ib                = (outp.data() == buff[0].data()) ? 0 : 1;
-        } else {
-            complexPreprocess(Ncvec, inputSpan, buff[ib], e_span);
-
-            std::span<T> outp = cfftf1_ps<Direction::Forward, T>(n_vecs_cplx, buff[ib], buff[0], buff[1], twiddle_span, setup._radixPlan);
-            ib                = (outp.data() == buff[0].data()) ? 0 : 1;
-
-            // interleave
-            T* RESTRICT pBuffer = std::assume_aligned<64>(buff[ib].data());
-            for (std::size_t k = 0; k < Ncvec; ++k) {
-                const std::size_t k2 = 2 * k * L;
-
-                V r(pBuffer + k2, stdx::vector_aligned);
-                V i(pBuffer + k2 + L, stdx::vector_aligned);
-                V v0, v1;
-                interleave(r, i, v0, v1);
-                store_unchecked(v0, pBuffer + k2, stdx::vector_aligned);
-                store_unchecked(v1, pBuffer + k2 + L, stdx::vector_aligned);
-            }
-        }
-    }
-
-    if (buff[ib].data() != outputSpan.data()) {
-        T* RESTRICT pBuffer = std::assume_aligned<64>(buff[ib].data());
-        T* RESTRICT pOutput = std::assume_aligned<64>(outputSpan.data());
-        // std::memcpy(pBuffer, pOutput, Ncvec * 2 * L * sizeof(T)); -- outch
-        std::memcpy(pOutput, pBuffer, Ncvec * 2 * L * sizeof(T));
-        ib = !ib;
-    }
-    assert(buff[ib].data() == outputSpan.data());
-}
-
-#ifdef ALWAYS_INLINE
-#undef ALWAYS_INLINE
-#endif
-#ifdef NEVER_INLINE
-#undef NEVER_INLINE
-#endif
-#ifdef RESTRICT
-#undef RESTRICT
-#endif
